+++ src/jobs/ProcessBulkOperationData.php
@@ -61,15 +61,15 @@ class ProcessBulkOperationData extends BaseBatchedJob
      */
     protected function loadData(): Batchable
     {
-        $downloadFile = false;
-
         // check to see if the file is still in temporary storage
-        if ($this->tempFilePath === null || !file_exists($this->tempFilePath)) {
+        if ($this->tempFilePath === null) {
+            // Make sure we only create one temp filename
             $this->tempFilePath = Assets::tempFilePath('jsonl');
-            $downloadFile = true;
         }
 
-        if ($downloadFile) {
+        // Only re-download if the file doesn't already exist
+        // If the queue is using multiple workers and the file leaks between them,
+        if (!file_exists($this->tempFilePath)) {
             // Retrieve remote file contents
             $client = Craft::createGuzzleClient();
             $response = $client->get($this->dataUrl);