+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Unreleased
 
+- Fixed a bug where file storage could be maxed out when using multiple queue workers.
 - Fixed a bug where contextual pricing countries weren’t being force to be capitalized.
 
 ## 6.1.0 - 2025-10-24